backup/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
stripe==8.0.0
python-dotenv==1.0.0
email-validator==2.3.0
